Refactor: put link related code in its own package
